requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
yfinance>=0.2.36  # Added for Yahoo Finance integration

# Trading & Analysis
alpaca-trade-api>=3.0.0
alpaca-py>=0.10.0  # Added for improved Alpaca SDK
ta-lib>=0.4.28; platform_system != "Darwin"  # Technical analysis library (optional)

# Testing
pytest>=8.0.0
pytest-mock>=3.12.0
pytest-asyncio>=0.23.5
pytest-cov>=4.1.0

# Development tools
black>=24.1.0  # Code formatting
flake8>=7.0.0  # Code linting
mypy>=1.8.0   # Type checking

# Logging & Time
pytz>=2024.1.0
python-dateutil>=2.8.2

# CLI & Data Processing
click>=8.1.7  # Added for CLI interface
scipy>=1.12.0  # Added for financial calculations
pandas-market-calendars>=4.3.1  # Added for market calendar 